server/requirements.txt
fastapi
uvicorn
duckdb
qdrant-client
sentence-transformers
anthropic
pandas
python-dotenv
requests
groq